refactor(config): isolate runtime admission and rate limit policy interpretation
